+++ NEWS.md
@@ -6,6 +6,8 @@
 * The MASS package is explicitly suggested to fix CRAN warnings on R-devel
   (#3657).
 
+* Set operations like `intersect()` and `setdiff()` reconstruct groups metadata (#3587).
+
 * Using namespaced calls to `base::sort()` and `base::unique()` from C++ code
   to avoid ambiguities when these functions are overridden (#3644).